[GOJHW-6]: implement grpc server and grpc client

+++ internal/dto/give.go
@@ -1,5 +1,5 @@
 package dto
 
 type GiveOrdersRequest struct {
-	Orders []uint `json:"orders"`
+	Orders []uint64 `json:"orders"`
 }
